so far working
basically, makes a new instance of bot. logs in, find the zau discord and the sync roles, finds the user, if in discord and dosent have role add it etc etc etc. has error handleing so dont need to worry bout that too. Bot logs in when api is started so it doesnt login and log off everytime a user connects